+++ tests/FSharp.ApiDocs.Tests/files/FsLib2/Library2.fs
@@ -94,7 +94,7 @@ type Test_Issue472_T () =
   member x.MultArgTupled(arg: (int * int)) = ()
 
 type ITestInterface =
-  abstract Test : unit -> RazorEngine.Templating.IRazorEngineService
+  abstract Test : unit -> DotLiquid.Templating.IDotLiquidService
   abstract FixScript : string -> string
 
 /// Issue 201 docs